Rename and add scope: @mergesium/shell-quote
The purpose of the scope is to make two things clear:
 - the package is distinct from the unscoped release 'shell-quote'
 - the focus of the scope is quality and precision of results (ie,
   the core values of Mergesium)

The name "shell-quote" is kept to signify that the intent is to keep
this package backwards compatible, or to make it obvious through
versioning when differences occur.

For instance, one of the immediate plans is to change the output of the
"quote" function to make it more readable to human eyes. There are many
ways to quote text for shell purposes, but not all are of equal quality.

Eg: the text "example" can be quoted as:
 - example
 - "example"
 - 'example'
 - 'exam''ple'
 - 'e''x''a''m''p''l''e'
 - \e\x\a\m\p\l\e

In some cases, Jordan's fork outputs quoted strings which are not as
easy on the eye as can be, although they are technically valid
quotations. One of the first priorities of the @mergesium fork is to
make these quotation human-readable.

Why ? Sometimes, quoted text might need to be included in published
documentation; ie, it is human-visible. In those cases, it should be as
legible as possible, while being correct. A quoting like \e\x\a\m\p\l\e
might tempt the editor to fix up the quoting, and in doing so, the
meaning of the result might change, such that it would have a different
effect than the original, when executed in an actual shell.
Documentation would not be safe to cut-and-paste into shell prompts.